aider: Add a help option that shows all available options we have.
# Aider chat conversation:

USER: add a help option that shows all available options we have

ASSISTANT: To add a help option that shows all available options, we need to modify the `src/cmdOptions.js` file to add a new condition to check if the process arguments include `-h` or `--help`. If they do, we log the help information to the console.

Here are the changes:

**src/cmdOptions.js**
```javascript
src/cmdOptions.js
<<<<<<< HEAD
export function handleCmdOptions() {
  if (process.argv.includes('--version') || process.argv.includes('-v')) {
    console.log(VERSION);
    process.exit(0);
  } else if (process.argv.includes('--list') || process.argv.includes('-l')) {
    console.log('Available templates:');
    for (const template in templateRepos) {
      console.log(template);
    }
    process.exit(0);
  }
}
=======
export function handleCmdOptions() {
  if (process.argv.includes('--version') || process.argv.includes('-v')) {
    console.log(VERSION);
    process.exit(0);
  } else if (process.argv.includes('--list') || process.argv.includes('-l')) {
    console.log('Available templates:');
    for (const template in templateRepos) {
      console.log(template);
    }
    process.exit(0);
  } else if (process.argv.includes('--help') || process.argv.includes('-h')) {
    console.log(`
      Usage: nobj [options]

      Options:
        -v, --version       output the version number
        -l, --list          list all available templates
        -h, --help          display help for command
    `);
    process.exit(0);
  }
}
>>>>>>> updated
```

This change will add a help option that shows all available options when running the command with `-h` or `--help`.