aider: Modified `projectType.js` to set the project type to the first available type if it's not provided in the command line arguments.
# Aider chat conversation:

USER: It's working for one argument (`nobj test1`) but not for two (`nobj test33 vanilla`). In the second case, i still got asked for the template type

ASSISTANT: I see. It seems like the issue is with the way we're checking for the project type in the command line arguments. Currently, we're checking if the project type exists and is included in the project types. However, if the project type is not provided, we're not skipping the question. We need to modify the `projectType.js` file to handle this case.

Here's what we need to do:

1. Modify the `projectType.js` file to check if the project type is provided in the command line arguments and if it's valid. If it's not provided or not valid, we should skip the question.

Here's the *edit block* for the change:

```javascript
src/projectType.js
<<<<<<< HEAD
  const projectType = process.argv[3];
  if (projectType && projectTypes.includes(projectType)) {
    settings.projectType = projectType;
  } else {
    const answer = await inquirer.prompt({
      name: 'project_type',
      type: 'list',
      message: 'Choose your template: ',
      choices: projectTypes,
    });
    settings.projectType = answer.project_type;
  }
=======
  const projectType = process.argv[3];
  if (projectType && projectTypes.includes(projectType)) {
    settings.projectType = projectType;
  } else if (!projectType) {
    settings.projectType = projectTypes[0];
  } else {
    const answer = await inquirer.prompt({
      name: 'project_type',
      type: 'list',
      message: 'Choose your template: ',
      choices: projectTypes,
    });
    settings.projectType = answer.project_type;
  }
>>>>>>> updated
```

This change will set the project type to the first available type if it's not provided in the command line arguments. If it's provided but not valid, it will still ask the question.